requirements-api.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.54.0
supabase==2.9.1
numpy==1.26.4
